+++ CHANGELOG.md
@@ -56,6 +56,8 @@
   * Removed groupId requirement in `ConvertTo-IntuneMobileAppAssignment`.
 * MISC
   * Centralized more resource testing to the testing function.
+  * Improved support for `ErrorAction` during export.
+    FIXES [#6610](https://github.com/microsoft/Microsoft365DSC/issues/6610)
 * DEPENDENCIES
   * Updated MSCloudLoginAssistant to version 1.1.53